refactor: 코드 패키지로 묶어서 정리
