+++ src/main/java/NotModified304/Scatch/dto/timeTable/ttc/TimeTableWithCourseRequestDto.java
@@ -1,7 +1,7 @@
-package NotModified304.Scatch.dto.timeTableWithCourse;
+package NotModified304.Scatch.dto.timeTable.ttc;
 
-import NotModified304.Scatch.dto.course.CourseRequestDto;
-import NotModified304.Scatch.dto.timeTableDetail.TimeTableDetailRequestDto;
+import NotModified304.Scatch.dto.timeTable.course.CourseRequestDto;
+import NotModified304.Scatch.dto.timeTable.ttd.TimeTableDetailRequestDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
