+++ src/main/java/NotModified304/Scatch/dto/timeTable/ttc/TimeTableWithCourseUpdateDto.java
@@ -1,7 +1,7 @@
-package NotModified304.Scatch.dto.timeTableWithCourse;
+package NotModified304.Scatch.dto.timeTable.ttc;
 
-import NotModified304.Scatch.dto.course.CourseUpdateDto;
-import NotModified304.Scatch.dto.timeTableDetail.TimeTableDetailUpdateDto;
+import NotModified304.Scatch.dto.timeTable.course.CourseUpdateDto;
+import NotModified304.Scatch.dto.timeTable.ttd.TimeTableDetailUpdateDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
