+++ src/main/java/NotModified304/Scatch/dto/timeTable/ttd/TimeTableDetailRequestDto.java
@@ -1,5 +1,4 @@
-package NotModified304.Scatch.dto.timeTableDetail;
-
+package NotModified304.Scatch.dto.timeTable.ttd;
 
 import lombok.*;
 
@@ -7,12 +6,13 @@ import java.time.LocalTime;
 
 @Getter
 @Setter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class TimeTableDetailUpdateDto {
-    private Long timeTableDetailId;
-    private Integer weekday;
+@Builder
+public class TimeTableDetailRequestDto {
+    private Long timeTableId;
+    private Long courseId;
+    private int weekday;
     private String location;
     private LocalTime startTime;
     private LocalTime endTime;
