+++ src/main/java/NotModified304/Scatch/service/CourseService.java
@@ -1,8 +1,9 @@
 package NotModified304.Scatch.service;
 
 import NotModified304.Scatch.domain.Course;
-import NotModified304.Scatch.dto.course.CourseRequestDto;
-import NotModified304.Scatch.dto.course.CourseUpdateDto;
+import NotModified304.Scatch.dto.timeTable.course.CourseRequestDto;
+import NotModified304.Scatch.dto.timeTable.course.CourseUpdateDto;
+import NotModified304.Scatch.repository.interfaces.AssignmentRepository;
 import NotModified304.Scatch.repository.interfaces.CourseRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CourseService {
     private final CourseRepository courseRepository;
+    private final AssignmentRepository assignmentRepository;
 
     public Course findCourse(Long id) {
         return courseRepository.findById(id).orElseThrow(
@@ -40,9 +42,19 @@ public class CourseService {
         String newInstructor = dto.getInstructor();
         String newColor = dto.getColor();
         
-        if(newTitle != null) course.setTitle(newTitle);
+        if(newTitle != null) {
+            course.setTitle(newTitle);
+
+            // 해당 강좌의 과제에서도 title 업데이트
+            assignmentRepository.updateCourseTitle(course.getId(), newTitle);
+        }
         if(newInstructor != null) course.setInstructor(newInstructor);
-        if(newColor != null) course.setColor(newColor);
+        if(newColor != null) {
+            course.setColor(newColor);
+
+            // 해당 강좌의 과제 색 업데이트
+            assignmentRepository.updateColor(course.getId(), newColor);
+        }
     }
 
     // 강좌 정보 삭제
