+++ src/main/java/NotModified304/Scatch/service/TimeTableDetailService.java
@@ -3,12 +3,9 @@ package NotModified304.Scatch.service;
 import NotModified304.Scatch.domain.Course;
 import NotModified304.Scatch.domain.TimeTable;
 import NotModified304.Scatch.domain.TimeTableDetail;
-import NotModified304.Scatch.dto.course.CourseUpdateDto;
-import NotModified304.Scatch.dto.timeTable.TimeTableResponseDto;
-import NotModified304.Scatch.dto.timeTableDetail.TimeTableDetailRequestDto;
-import NotModified304.Scatch.dto.timeTableDetail.TimeTableDetailResponseDto;
-import NotModified304.Scatch.dto.timeTableDetail.TimeTableDetailUpdateDto;
-import NotModified304.Scatch.dto.timeTableWithCourse.TimeTableWithCourseResponseDto;
+import NotModified304.Scatch.dto.timeTable.ttd.TimeTableDetailRequestDto;
+import NotModified304.Scatch.dto.timeTable.ttd.TimeTableDetailUpdateDto;
+import NotModified304.Scatch.dto.timeTable.ttc.TimeTableWithCourseResponseDto;
 import NotModified304.Scatch.repository.interfaces.AssignmentRepository;
 import NotModified304.Scatch.repository.interfaces.CourseRepository;
 import NotModified304.Scatch.repository.interfaces.TimeTableDetailRepository;
