+++ src/main/java/NotModified304/Scatch/service/TimeTableService.java
@@ -3,9 +3,9 @@ package NotModified304.Scatch.service;
 import NotModified304.Scatch.domain.Course;
 import NotModified304.Scatch.domain.TimeTable;
 import NotModified304.Scatch.domain.TimeTableDetail;
-import NotModified304.Scatch.dto.timeTable.TimeTableRequestDto;
-import NotModified304.Scatch.dto.timeTable.TimeTableResponseDto;
-import NotModified304.Scatch.dto.timeTable.TimeTableUpdateDto;
+import NotModified304.Scatch.dto.timeTable.tt.TimeTableRequestDto;
+import NotModified304.Scatch.dto.timeTable.tt.TimeTableResponseDto;
+import NotModified304.Scatch.dto.timeTable.tt.TimeTableUpdateDto;
 import NotModified304.Scatch.repository.interfaces.AssignmentRepository;
 import NotModified304.Scatch.repository.interfaces.CourseRepository;
 import NotModified304.Scatch.repository.interfaces.TimeTableDetailRepository;
